Implement the necessary transformers for TLS support.
TLS has its own termination state which complicates the termination
of a Flow. Blocking the drain operation is used to make sure that termination
is sequenced properly by Transports. This allows TLS to be implemented without
isolating it into a separate Flow. (Which is a complication of a the Sector instead of the Flow).

Transports provide a solution for a stack of layers where each layer is terminated according
to their position on the stack. Normally, layers manage their own termination state and
Transports responds to changes in that state by popping the layer off of the stack.