Fix drain logic to manage Inexorable obstructions.
Make Inexorable obstructions trigger drained() operations so
that termination will not be blocked by a Flow that cannot process anyways.

Arguably, the operation should be sensitive to the placement of the Transformer
that is obstructing, but the idea is that the higher level will not be able to perform
its task regardless of the positioning of an obstruction allowing us to complete the
drain and signal the opposite to finish its drain as well.