Rework scheduler to not be in the processor set.
Scheduler() is actually a supporting resource and will likely lose Processor() status
as it should not be blocking Sector exits. It's generally treated specially as termination
procedures may depend on it working, so it's important that it's not treated as a regular
processor anyways.

This with updates to bin.cache allow exits to occur naturally again.