Refactor fault.io away from Transformer based Flows. [Step 1/2]
[ Flows ]

By Transformer and Flow in the initial line, the concept as defined
in the original being patched. Essentially, this change merges the two
concepts into just a "Flow".

The common issue being that Transformers needed their own termination
state and access to the flow's obstruction state. By making the transformers
flows, the problem of constant controller access is resolved. Additionally,
original Flows (now called Transformations) has some Sector qualities with
respect to the containment of other Processor subclasses (Transformer).

The original Flow concept now exists as "Transformation", and some Transformer
subclasses still exists as their conversion meant dealing with the drain concept
introduced to properly sequence termination of Transport's and KernelPort's.

The completed refactoring will eliminate Transformer and Transformation entirely.
Transports will become its own Flow subclass managing the I/O layers of a connection
and likely integrating the Kernel layer as a stack entry.

[ Subflows ]

The original Sequencing and Distribution have been renamed to
Catenation and Division, and the protocol for managing subflows is
now handled by &libio.FlowControl events rather than direct assignment of a callback.

Arguably, the direct callbacks reduced the number of hash resolutions,
but the increased in complexity and difficulty in testing. The performance
gains were likely too mild to justify, and the new design allowed the QueueProtocol
to be eliminated as well.

[ Interfaces and Sectors ]

Sector subclasses are a bad idea, and the use along with the Nothing processor
pointed the way to have the (Interface) subclass replace Nothing.

It's nearly 6-of-1, but with the exception being that descendants can access
the Interface portions without causing a circular reference. The convenience
of accessing the interfaces from .controller properties is nearly maintained
by having all Interface subclasses use the same sector placement.