Relocate the high-level HTTP support into web.
This means that libdaemon and libroot depends on web for the time being.
Eventually, their Host definition may be simplified to the point that
web.libhttpd features may be unnecessary.

Notably, this sets a clear line for what belongs inside a protocol implementation
module and what will likely become a model for other implementations. A low-level adapter
that hands off work to the higher level routers/managers.

For libio, this will probably lead the way to managing a registry for selected protocol
implementations. Rather than refer to io.http directly, a requested protocol will be
mapped to an implementation module causing its import.