Split invocation from sectors configuration and modify API.
The invocation XML grammar defined by fault.system is much more flexible, so refine
sectors configuration to be primarily about port binding.

rootd was also modified. The set of daemons is managed inside a directory inside the instance
directory. This allows for faultd instances to be nested inside each other.