Introduce `io` module and perform relevant refactoring.
- Replace `ctx_enqueue_task` with `critical` method.
- Refactor Catenation.f_transfer to allow for transparent mitres.
- Eliminate Inlet in favor of a generalized Relay.
- Introduce io.Transport and io.Transfer as the appropriate containers for flows.
- Eliminate security module and temporarily keep the operations function in `io`.
- Correct a defect in Division that conflated subflow termination that resulted in ambiguity.
- Various bitrot removed.

Primarily, this is the introduction of the `io` module implementing the primary containers
for flows: Transport and Transfer. While, these interfaces are soon to change, the working
state and the basic tests encouraged a commit.

There is likely some bitrot involving system.io channel allocations, but the primarily
one used by clients has been adjusted.