Relocate Mitre from fault.web.
This generalized form places much responsibility downstream of the mitre, especially for
clients. However, the prior presumption made from the implementation in fault.web drew
conclusions about how clients would connect requests to responses. A reasonable
implementation, but potentially undesireable.

The latest implementation covers clients and servers by providing two interfaces
for retrieving the set of protocol transactions. One for servers which need
to have the output slot reserved and one for clients that presumes that
that has already happened.