Use system.process.scheduler across Process instances and simplify Matrix.
Usage of the single process scheduler is the beginning of eliminating
the system.Process class entirely from kernel. system.Context inheriting
the IO instance and Fabric eventually being eliminated or partially
integrated as well.

Matrix refactoring intends to bind the purpose of the implementation to
&.kernel. Route and the distinct I/O regions (id parameters) being
removed as consolidating at the Matrix instance appears to be conflating
concerns. I/O "regions" can be handled with multiple instances, and
routes with, likely, distinct classes.